Markdown File Version 1.0
This is obviously not the perfect version. I did it to show you we have
some graph problem(some of them are not showing properly except the map
graphs). I would suggest everyone take a look at it and see if you can
revise your graph code). Besides, we need to do the structure better
after finishing the story telling part.